+++ src/Bicep.Core/TypeSystem/BooleanLiteralType.cs
@@ -1,6 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+using System;
+
 namespace Bicep.Core.TypeSystem;
 
 public class BooleanLiteralType : TypeSymbol
@@ -22,11 +24,7 @@ public class BooleanLiteralType : TypeSymbol
     public bool Value { get; }
 
     public override bool Equals(object? other) =>
-        other is BooleanLiteralType otherBoolLiteral ? otherBoolLiteral == this : false;
+        other is BooleanLiteralType otherBoolLiteral ? Value == otherBoolLiteral.Value : false;
 
-    public override int GetHashCode() => (GetType(), Value).GetHashCode();
-
-    public static bool operator ==(BooleanLiteralType? a, BooleanLiteralType? b) => a?.Value == b?.Value;
-
-    public static bool operator !=(BooleanLiteralType? a, BooleanLiteralType? b) => a?.Value != b?.Value;
+    public override int GetHashCode() => HashCode.Combine(TypeKind, Value);
 }
